backend/requirements.txt
# Core backend
Flask==3.1.0
flask-cors==6.0.1
requests==2.32.5

# File processing
pdfplumber==0.11.7
python-docx==1.2.0

# Plagiarism API (optional, for Copyleaks)
copyleaks==5.1.0

# Machine learning / semantic similarity (optional)
torch>=2.1.0
sentence-transformers>=2.2.2

# Utilities
python-dotenv==1.1.1
